% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_emissions_by.R
\name{annual_emissions_by}
\alias{annual_emissions_by}
\title{Total emissions by year}
\usage{
annual_emissions_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\description{
Total \code{ems_qty} by \code{year}, ..., \code{ems_unit}
}
\examples{
df <- data_frame(year = 1990:1993, foo = rep(c("bar", "baz"), 2), ems_qty = 1:4, ems_unit = "tons/yr")
df \%>\% annual_emissions_by(foo, verbose = TRUE)

}
\seealso{
\code{\link[=annual_throughputs_by]{annual_throughputs_by()}}

\code{\link[=tabulate_emissions_by]{tabulate_emissions_by()}}

\code{\link[=total_emissions_by]{total_emissions_by()}}
}
