% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_quantities_by.R
\name{annual_quantities_by0}
\alias{annual_quantities_by0}
\title{Total quantities by year}
\usage{
annual_quantities_by0(input_data, ..., qty_var, unit_var, year_var = "year",
  digits = Inf, signif = Inf)
}
\description{
Total quantities by year
}
