% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_throughputs_by.R
\name{annual_throughputs_by}
\alias{annual_throughputs_by}
\title{Total throughputs by year}
\usage{
annual_throughputs_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\description{
Total \code{tput_qty} by \code{year}, ..., \code{tput_unit}
}
\seealso{
\code{\link[=annual_emissions_by]{annual_emissions_by()}}

\code{\link[=tabulate_throughputs_by]{tabulate_throughputs_by()}}

\code{\link[=total_throughputs_by]{total_throughputs_by()}}
}
