% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scalars.R
\name{apply_scalars}
\alias{apply_scalars}
\title{Multiply a column in one dataset by one or more variables from another dataset}
\usage{
apply_scalars(input_data, using, value_col, key_col = NULL,
  join_na = TRUE, keep_as = NULL, verbose = FALSE, warn = TRUE)
}
\description{
Multiply a column in one dataset by one or more variables from another dataset
}
\examples{
demo("apply_scalars", package = "inventory")

}
