% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Like \code{round()}, but you're not limited to 0.1s, 0.01s, etc.}
\usage{
bin(x, by = 1)
}
\description{
Like \code{round()}, but you're not limited to 0.1s, 0.01s, etc.
}
\examples{
x <- 0:8
bin(x, by = 3)

}
