% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_forward.R
\name{carry_forward}
\alias{carry_forward}
\title{Carry non-NA values forward, replacing NAs}
\usage{
carry_forward(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Carry non-NA values forward, replacing NAs
}
\note{
Similar to \code{na.locf} in package \code{zoo}
}
\examples{
  carry_forward(c(1, NA, 2, 3, NA, 5, 6, NA))
  carry_forward(c(NA, NA, NA, 1, NA, 2, 3, NA, 5, 6, NA))

}
