% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_qty.R
\name{convert_quantities}
\alias{convert_quantities}
\alias{converty_qty}
\title{Easily convert between different units of measurement.}
\usage{
convert_quantities(x, from, to)
}
\description{
Generic function to convert a numeric vector from one scale of
measurement to another.
}
\details{
Relies on the nifty \code{units} package.
}
\examples{
convert_quantities(1, "ton", "lbs") # mass
convert_quantities(100, "tons/day", "lbs/yr") # mass rate
convert_quantities(1, "yard^3", "ft^3") # volume
convert_quantities(212, "degF", "degC") # temperature

}
\seealso{
\itemize{
\item \code{vignette("units", package = "units")}
}
}
