% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{int}
\alias{int}
\title{Safely coerce to integer(s)}
\usage{
int(...)
}
\arguments{
\item{x}{character or factor}
}
\value{
integer
}
\description{
Safely coerce to integer(s)
}
\examples{
int(1, 3:5)
int(as.factor(c(foo = 2, bar = 99)))
int(ALA = 1, CC = 2)
int(1:5)

}
