% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_int.R
\name{is_int}
\alias{is_int}
\title{Test for integer-ish-ness}
\usage{
is_int(x, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{numeric or integer}

\item{tolerance}{passed to \code{all.equal}}
}
\description{
Is \verb{x`` equivalent to the integer (rounded) version of }x`? Works for doubles, etc.
}
\note{
Probably someone else has implemented a better version
}
