% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scalar.R
\name{is_scalar}
\alias{is_scalar}
\title{Test whether a vector has exactly one element}
\usage{
is_scalar(x, metric = c("length", "elements"), .xname = get_name_in_parent(x))
}
\description{
imported from package \code{assertive}
}
