% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lerp.R
\name{lerp}
\alias{lerp}
\title{Linear Interpolation}
\usage{
lerp(x, ...)
}
\arguments{
\item{...}{further arguments}

\item{object}{vector or data frame}
}
\value{
modified copy, with NAs replaced by linearly interpolated values
}
\description{
Linear Interpolation
}
\examples{
year <- c(1991, 1993, 1995, 1998)
sales <- c(21, NA, 40, 45)
lerp(year, sales)
sales_model <- lerp(sales ~ year, data = data.frame(year, sales))
predict(sales_model, data_frame(year = 1990:2000))
predict(sales_model, data_frame(year = 1990:2000), extrapolate = TRUE)
}
