% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locf.R
\name{locf}
\alias{locf}
\title{Last Observation Carried Forward}
\usage{
locf(object, ...)
}
\arguments{
\item{object}{vector or data frame}

\item{...}{further arguments}
}
\value{
modified copy, with NAs replaced by last observations carried forward
}
\description{
Last Observation Carried Forward
}
\examples{
sales <- c(21, NA, 40, 45)
locf(sales)
year <- c(1991, 1993, 1995, 1998)
mod <- locf(sales ~ year, data = data.frame(year, sales))
predict(mod, data_frame(year = 1994:2000))
}
