% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_each.R
\name{multiply_each}
\alias{multiply_each}
\title{Multiply numbers by a given amount}
\usage{
multiply_each(input_data, ..., by = 1)
}
\arguments{
\item{input_data}{tbl containing columns to total}

\item{\dots}{additional parameters passed on to methods}
}
\description{
Multiply numbers by a given amount
}
