% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct-class.R
\name{percent}
\alias{percent}
\alias{as_percent}
\alias{is_percent}
\title{\code{percent} vector}
\usage{
percent(x = double())

as_percent(x)

is_percent(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{percent()}: A numeric vector
\item For \code{is_percent()}: An object to test.
}}
}
\value{
An S3 vector of class \code{qtytools_pct}.
}
\description{
This creates a double vector that represents percentages so when it is
printed, it is multiplied by 100 and suffixed with \verb{\%}.
}
\examples{
percent(c(0.25, 0.5, 0.75))
}
