% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_of.R
\name{percent_of}
\alias{percent_of}
\title{share_of}
\usage{
percent_of(x)
}
\arguments{
\item{x}{(numeric) vector with no \code{NA}s}
}
\description{
Readable shorthand for \code{x / sum(x)}. Throws an error if \code{x} contains \code{NA}.
}
\seealso{
\itemize{
\item \code{\link[=share_of]{share_of()}}
}
}
