% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct-class.R
\name{qtytools-vctrs}
\alias{qtytools-vctrs}
\alias{vec_ptype2.qtytools_pct}
\alias{vec_cast.qtytools_pct}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{qtytools_pct}(x, y, ...)

\method{vec_cast}{qtytools_pct}(x, to, ...)
}
\description{
Internal vctrs methods
}
\note{
Our percent class is coercible back and forth with double vectors.

The default method provides a user friendly error message if the coercion
doesn’t exist and makes sure \code{NA} is handled in a standard way. \code{NA} is
technically a logical vector, but we want to stand in for a missing value of
any type.
}
\keyword{internal}
