% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_half_up.R
\name{round_half_up}
\alias{round_half_up}
\title{round_half_up}
\usage{
round_half_up(x, digits = 0)
}
\arguments{
\item{x}{numeric}

\item{digits}{integer}
}
\description{
Round according to human-friendly "round-half-up" convention.
}
\details{
R's \code{round()} follows the IEEE "round-to-even" rule, meaning that 1.5 will be rounded to 2, but 2.5 will also be rounded to 2. This isn't what many (most?) humans expect.
}
