% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_of.R
\name{share_of}
\alias{share_of}
\title{share_of}
\usage{
share_of(x)
}
\arguments{
\item{x}{(numeric) vector with no \code{NA}s}
}
\description{
Readable shorthand for \code{x / sum(x)}. Throws an error if \code{x} contains \code{NA}.
}
\seealso{
\itemize{
\item \code{\link[=percent_of]{percent_of()}}
}
}
