% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_quantities_by.R
\name{sum_quantities_by}
\alias{sum_quantities_by}
\title{Sum Quantities In Tabular Data}
\usage{
sum_quantities_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\arguments{
\item{input_data}{(tabular)}

\item{...}{variables to group by}

\item{digits}{(integer) passed to \code{\link[=total]{total()}}}

\item{signif}{(integer) passed to \code{\link[=total]{total()}}}

\item{verbose}{(logical)}
}
\description{
All of the variables in \code{input_data} whose names end in \code{_qty} will be summed. Corresponding variables whose names end in \code{_unit} will be preserved. Ignores \code{NA}s.
}
\examples{
df <- tibble(year = 1990:1993, foo = rep(c("bar", "baz"), 2), ems_qty = 1:4, ems_unit = "tons/yr")
df \%>\% sum_quantities_by(foo, verbose = TRUE)
df \%>\% sum_quantities_by("year", foo, verbose = TRUE)
df \%>\% sum_quantities_by("year", foo, digits = 2, verbose = TRUE)

}
\seealso{
Other sum_by: \code{\link{sum_emissions_by}},
  \code{\link{sum_throughputs_by}}
}
\concept{sum_by}
