% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_throughputs_by.R,
%   R/sum_annual_throughputs_by.R
\name{sum_throughputs_by}
\alias{sum_throughputs_by}
\alias{sum_annual_throughputs_by}
\alias{annual_throughputs_by}
\title{Sum Throughputs In Tabular Data}
\usage{
sum_throughputs_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))

sum_annual_throughputs_by(...)
}
\arguments{
\item{input_data}{(tabular)}

\item{...}{variables to group by}

\item{digits}{(integer) passed to \code{\link[=total]{total()}}}

\item{signif}{(integer) passed to \code{\link[=total]{total()}}}

\item{verbose}{(logical)}
}
\description{
All of the variables in \code{input_data} whose names end in \code{tput_qty} will be summed. \code{tput_unit} will be preserved. Ignores \code{NA}s.

\code{sum_annual_throughputs_by()} is a specialized variant that always groups by \code{year}.
}
\note{
\code{annual_throughputs_by(...)} is equivalent to \code{sum_annual_throughputs_by()}, but the latter is now preferred.
}
\examples{
df <- tibble(year = 1990:1993, foo = rep(c("bar", "baz"), 2), tput_qty = 1:4, tput_unit = "kgal")
df \%>\% sum_annual_throughputs_by(foo, verbose = TRUE)

}
\seealso{
Other sum_by: \code{\link{sum_emissions_by}},
  \code{\link{sum_quantities_by}}
}
\concept{sum_by}
