% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_emissions_by.R,
%   R/tabulate_quantities_by.R, R/tabulate_throughputs_by.R
\name{tabulate_emissions_by}
\alias{tabulate_emissions_by}
\alias{tabulate_quantities_by}
\alias{tabulate_throughputs_by}
\title{Total, then spread, annual quantities}
\usage{
tabulate_emissions_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))

tabulate_quantities_by(input_data, ..., fill = 0, digits = Inf,
  signif = Inf, verbose = getOption("verbose"))

tabulate_throughputs_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\arguments{
\item{input_data}{(tabular) must have a column ending in \code{_qty}}

\item{...}{variables to group and spread by (see Details).}

\item{digits}{(optional) round results to this many digits}

\item{signif}{(optional) round results to this many \emph{significant} digits}

\item{verbose}{(logical)}

\item{fill}{value for "empty cells"; see \link[tidyr:spread]{spread()}}
}
\description{
Total, then spread, annual quantities
}
\details{
The last variable in \code{...} is used to spread (that is, to form new columns).
}
\section{Functions}{
\itemize{
\item \code{tabulate_emissions_by}: Tabulate \code{ems_qty}

\item \code{tabulate_throughputs_by}: Tabulate \code{tput_qty}
}}

