% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_q.R
\name{top_q}
\alias{top_q}
\title{Select top q rows (by value)}
\usage{
top_q(x, q, wt, na.rm = TRUE, .arrange = FALSE,
  verbose = getOption("verbose"))
}
\arguments{
\item{x}{(tabular data)}

\item{q}{(numeric) between 0 and 1; passed to \code{\link[=quantile]{quantile()}}}

\item{wt}{the variable to use for ordering.}

\item{na.rm}{(logical) ignore \code{NA}s in \code{wt}?}

\item{.arrange}{(logical) sort the results by \code{wt}?}
}
\description{
Select top q rows (by value)
}
\examples{
top_q(mtcars, q = 0.9, wt = mpg)

}
