% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\title{Add numbers, ignoring NAs}
\usage{
total(x, na.rm = TRUE, digits = Inf, signif = Inf)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{for \code{round}}

\item{signif}{for \code{signif}}
}
\value{
sum of (non-missing) values in \code{x}
}
\description{
Add numbers, ignoring NAs
}
