% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_each.R
\name{total_each}
\alias{total_each}
\title{Total selected columns of a data frame}
\usage{
total_each(input_data, ..., digits = Inf, signif = Inf)
}
\arguments{
\item{input_data}{input_data data frame}

\item{\dots}{bare names of variables (columns) to total}
}
\description{
Total selected columns of a data frame
}
\note{
Typically used in conjunction with \code{group_by}
}
