% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_emissions_by.R
\name{total_emissions_by}
\alias{total_emissions_by}
\title{Total emissions, ignoring NAs}
\usage{
total_emissions_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\description{
Total \code{ems_qty} by ...
}
\note{
Ignores \code{year} and \code{pol_xxx}, unless you say otherwise!
}
\seealso{
\code{\link[=tabulate_throughputs_by]{tabulate_throughputs_by()}}

\code{\link[=annual_emissions_by]{annual_emissions_by()}}
}
