% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_quantities_by.R, R/total_quantities_by.R
\name{annual_quantities_by}
\alias{annual_quantities_by}
\alias{total_quantities_by}
\title{Sum up quantities in tabular data}
\usage{
annual_quantities_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))

total_quantities_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\arguments{
\item{input_data}{(tabular)}

\item{...}{variables to group by}

\item{digits}{(integer) passed to \code{\link[=total]{total()}}}

\item{signif}{(integer) passed to \code{\link[=total]{total()}}}

\item{verbose}{(logical)}
}
\description{
Sum, ignoring \code{NA}s, all of the variables in \code{input_data} whose names end in \code{_qty}.
}
\section{Functions}{
\itemize{
\item \code{annual_quantities_by}: Totals by `year``
}}

\examples{
df <- data_frame(year = 1990:1993, foo = rep(c("bar", "baz"), 2), ems_qty = 1:4, ems_unit = "tons/yr")
df \%>\% annual_quantities_by(foo, verbose = TRUE)

df <- data_frame(year = 1990:1993, foo = rep(c("bar", "baz"), 2), ems_qty = 1:4, ems_unit = "tons/yr")
df \%>\% total_quantities_by(foo, verbose = TRUE)
df \%>\% total_quantities_by("year", foo, verbose = TRUE)
df \%>\% total_quantities_by("year", foo, digits = 2, verbose = TRUE)

}
\seealso{
\code{\link[=total]{total()}}
}
