% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_throughputs_by.R
\name{total_throughputs_by}
\alias{total_throughputs_by}
\title{Total throughputs, ignoring NAs}
\usage{
total_throughputs_by(input_data, ..., digits = Inf, signif = Inf,
  verbose = getOption("verbose"))
}
\description{
Total \code{tput_qty} by ...
}
\note{
Ignores \code{year} unless you say otherwise!
}
\seealso{
[tabulate_throughputs_by()

\code{\link[=annual_throughputs_by]{annual_throughputs_by()}}
}
