% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lerp.R
\name{lerp}
\alias{lerp}
\alias{lerp.default}
\alias{predict.lerp}
\title{Linear interpolation}
\usage{
lerp(object, ...)

\method{lerp}{default}(
  object,
  x = NULL,
  xout = NULL,
  nm = FALSE,
  extrapolate = FALSE,
  ...,
  verbose = getOption("verbose")
)

\method{predict}{lerp}(
  object,
  envir = NULL,
  method = "linear",
  extrapolate = FALSE,
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{object}{(see Details) must contain values to be interpolated}

\item{...}{further arguments passed to \code{\link[stats:approxfun]{stats::approx()}}}

\item{x}{(vector) x-coords; if not supplied, values to be interpolated are assumed to be evenly spaced}

\item{xout}{(numeric) points at which to make predictions}

\item{verbose}{(logical) display messages}
}
\description{
Linear interpolation
}
\details{
Names may be assigned to the resulting vector. If \code{nm} is \code{FALSE}, then no names will be assigned.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: replace NAs with interpolated values

\item \code{lerp}: calculate values using a formula
}}

\examples{
year <- c(1991, 1993, 1995, 1998)
sales <- c(21, NA, 40, 45)
sales_model <- lerp(sales ~ year, data = data.frame(year, sales))

lerp(sales)
lerp(sales, year)

predict(sales_model, tibble::tibble(year = 1990:2000))
predict(sales_model, tibble::tibble(year = 1990:2000), extrapolate = TRUE)

}
\concept{lerp}
