% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_quantities_by.R,
%   R/sum_annual_quantities_by.R
\name{sum_quantities_by}
\alias{sum_quantities_by}
\alias{sum_annual_quantities_by}
\alias{annual_quantities_by}
\title{Sum Quantities In Tabular Data}
\usage{
sum_quantities_by(
  input_data,
  ...,
  digits = Inf,
  signif = Inf,
  verbose = getOption("verbose")
)

sum_annual_quantities_by(...)
}
\arguments{
\item{input_data}{(tabular)}

\item{...}{variables to group by}

\item{digits}{(integer) passed to \code{\link[=total]{total()}}}

\item{signif}{(integer) passed to \code{\link[=total]{total()}}}

\item{verbose}{(logical)}
}
\description{
All of the variables in \code{input_data} whose names end in \verb{_qty} will be summed. Corresponding variables whose names end in \verb{_unit} will be preserved. Ignores \code{NA}s.

\code{sum_annual_quantities_by()} is a specialized variant that always groups by \code{year}.
}
\note{
\code{annual_quantities_by(...)} is equivalent to \code{sum_annual_quantities_by()}, but the latter is now preferred.
}
\examples{
df <- tibble(year = 1990:1993, foo = rep(c("bar", "baz"), 2), ems_qty = 1:4, ems_unit = "ton/yr")
df \%>\% sum_quantities_by(foo, verbose = TRUE)
df \%>\% sum_quantities_by("year", foo, verbose = TRUE)
df \%>\% sum_quantities_by("year", foo, digits = 2, verbose = TRUE)

df <- tibble(year = 1990:1993, foo = rep(c("bar", "baz"), 2), ems_qty = 1:4, ems_unit = "ton/yr")
df \%>\% sum_annual_quantities_by(foo, verbose = TRUE)

}
\seealso{
Other sum_by: 
\code{\link{sum_emissions_by}()},
\code{\link{sum_throughputs_by}()}
}
\concept{sum_by}
